+++ test/data/init-script.sql
@@ -1,0 +1,6 @@
+
+create table customers (
+     id bigint not null,
+     name varchar not null,
+     primary key (id)
+);